📛rename: 🎨design:  ~Style 컴포넌트명 수정 및 스타일 수정
- ~Style -> Container
- 모바일 column 수정

커뮤니티 카드 컴포넌트 구현 #29
